requirements.txt
pytorch
numpy
scikit-learn
pandas